feat: update tabs to take menu render prop
