+++ libs/candles-chart/src/lib/candles-chart.tsx
@@ -1,15 +1,5 @@
 import 'pennant/dist/style.css';
-import {
-  CandlestickChart,
-  ChartType,
-  Interval,
-  Overlay,
-  Study,
-  chartTypeLabels,
-  intervalLabels,
-  overlayLabels,
-  studyLabels,
-} from 'pennant';
+import { CandlestickChart, ChartType, Interval, Overlay, Study } from 'pennant';
 import { VegaDataSource } from './data-source';
 import { useApolloClient } from '@apollo/client';
 import { useMemo } from 'react';
@@ -19,18 +9,6 @@ import {
   getValidItem,
   getValidSubset,
 } from '@vegaprotocol/react-helpers';
-import {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuItemIndicator,
-  DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuTrigger,
-  Icon,
-} from '@vegaprotocol/ui-toolkit';
-import type { IconName } from '@blueprintjs/icons';
-import { IconNames } from '@blueprintjs/icons';
 import { t } from '@vegaprotocol/i18n';
 import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
@@ -81,13 +59,6 @@ export const useCandlesChartSettings = create<
   )
 );
 
-const chartTypeIcon = new Map<ChartType, IconName>([
-  [ChartType.AREA, IconNames.TIMELINE_AREA_CHART],
-  [ChartType.CANDLE, IconNames.WATERFALL_CHART],
-  [ChartType.LINE, IconNames.TIMELINE_LINE_CHART],
-  [ChartType.OHLC, IconNames.WATERFALL_CHART],
-]);
-
 export type CandlesChartContainerProps = {
   marketId: string;
 };
@@ -130,130 +101,26 @@ export const CandlesChartContainer = ({
   }, [client, marketId, pubKey]);
 
   return (
-    <div className="h-full flex flex-col">
-      <div className="px-3 lg:px-4 py-2 flex flex-row flex-wrap gap-2 bg-vega-clight-700 dark:bg-vega-cdark-700">
-        <DropdownMenu
-          trigger={
-            <DropdownMenuTrigger>
-              {t(`Interval: ${intervalLabels[interval]}`)}
-            </DropdownMenuTrigger>
-          }
-        >
-          <DropdownMenuContent>
-            <DropdownMenuRadioGroup
-              value={interval}
-              onValueChange={(value) => {
-                settings.setInterval(value as Interval);
-              }}
-            >
-              {Object.values(Interval).map((timeInterval) => (
-                <DropdownMenuRadioItem
-                  key={timeInterval}
-                  inset
-                  value={timeInterval}
-                >
-                  {intervalLabels[timeInterval]}
-                  <DropdownMenuItemIndicator />
-                </DropdownMenuRadioItem>
-              ))}
-            </DropdownMenuRadioGroup>
-          </DropdownMenuContent>
-        </DropdownMenu>
-        <DropdownMenu
-          trigger={
-            <DropdownMenuTrigger>
-              <Icon name={chartTypeIcon.get(chartType) as IconName} />
-            </DropdownMenuTrigger>
-          }
-        >
-          <DropdownMenuContent>
-            <DropdownMenuRadioGroup
-              value={chartType}
-              onValueChange={(value) => {
-                settings.setType(value as ChartType);
-              }}
-            >
-              {Object.values(ChartType).map((type) => (
-                <DropdownMenuRadioItem key={type} inset value={type}>
-                  {chartTypeLabels[type]}
-                  <DropdownMenuItemIndicator />
-                </DropdownMenuRadioItem>
-              ))}
-            </DropdownMenuRadioGroup>
-          </DropdownMenuContent>
-        </DropdownMenu>
-        <DropdownMenu
-          trigger={<DropdownMenuTrigger>{t('Overlays')}</DropdownMenuTrigger>}
-        >
-          <DropdownMenuContent>
-            {Object.values(Overlay).map((overlay) => (
-              <DropdownMenuCheckboxItem
-                key={overlay}
-                checked={overlays.includes(overlay)}
-                onCheckedChange={() => {
-                  const newOverlays = [...overlays];
-                  const index = overlays.findIndex((item) => item === overlay);
-
-                  index !== -1
-                    ? newOverlays.splice(index, 1)
-                    : newOverlays.push(overlay);
-
-                  settings.setOverlays(newOverlays);
-                }}
-              >
-                {overlayLabels[overlay]}
-                <DropdownMenuItemIndicator />
-              </DropdownMenuCheckboxItem>
-            ))}
-          </DropdownMenuContent>
-        </DropdownMenu>
-        <DropdownMenu
-          trigger={<DropdownMenuTrigger>{t('Studies')}</DropdownMenuTrigger>}
-        >
-          <DropdownMenuContent>
-            {Object.values(Study).map((study) => (
-              <DropdownMenuCheckboxItem
-                key={study}
-                checked={studies.includes(study)}
-                onCheckedChange={() => {
-                  const newStudies = [...studies];
-                  const index = studies.findIndex((item) => item === study);
-
-                  index !== -1
-                    ? newStudies.splice(index, 1)
-                    : newStudies.push(study);
-
-                  settings.setStudies(newStudies);
-                }}
-              >
-                {studyLabels[study]}
-                <DropdownMenuItemIndicator />
-              </DropdownMenuCheckboxItem>
-            ))}
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </div>
-      <div className="flex-1">
-        <CandlestickChart
-          dataSource={dataSource}
-          options={{
-            chartType: chartType,
-            overlays: overlays,
-            studies: studies,
-            notEnoughDataText: (
-              <span className="text-xs text-center">{t('No data')}</span>
-            ),
-          }}
-          interval={interval}
-          theme={theme}
-          onOptionsChanged={(options) => {
-            settings.merge({
-              overlays: options.overlays,
-              studies: options.studies,
-            });
-          }}
-        />
-      </div>
+    <div className="h-full">
+      <CandlestickChart
+        dataSource={dataSource}
+        options={{
+          chartType: chartType,
+          overlays: overlays,
+          studies: studies,
+          notEnoughDataText: (
+            <span className="text-xs text-center">{t('No data')}</span>
+          ),
+        }}
+        interval={interval}
+        theme={theme}
+        onOptionsChanged={(options) => {
+          settings.merge({
+            overlays: options.overlays,
+            studies: options.studies,
+          });
+        }}
+      />
     </div>
   );
 };
