+++ libs/candles-chart/src/lib/index.ts
@@ -1,4 +1,5 @@
 export * from './__generated__/Candles';
 export * from './__generated__/Chart';
 export * from './candles-chart';
+export * from './candles-menu';
 export * from './data-source';
